+++ test/json/replacer.test.js
@@ -65,12 +65,19 @@ describe('replacer', function () {
     assert.deepEqual(JSON.stringify(m), json);
   });
   
-  /*it('should stringify a Matrix, ccs storage format', function () {
+  it('should stringify a Matrix, ccs storage format', function () {
     var m = math.matrix([[1,2],[3,4]], 'ccs');
     var json = '{"mathjs":"CcsMatrix","values":[1,3,2,4],"index":[0,1,0,1],"ptr":[0,2,4],"size":[2,2]}';
 
     assert.deepEqual(JSON.stringify(m), json);
-  });*/
+  });
+
+  it('should stringify a Matrix, crs storage format', function () {
+    var m = math.matrix([[1,2],[3,4]], 'crs');
+    var json = '{"mathjs":"CrsMatrix","values":[1,2,3,4],"index":[0,1,0,1],"ptr":[0,2,4],"size":[2,2]}';
+
+    assert.deepEqual(JSON.stringify(m), json);
+  });
 
   it('should stringify a ResultSet', function () {
     var r = new ResultSet([1,2,new Complex(3,4)]);
@@ -86,13 +93,13 @@ describe('replacer', function () {
     assert.deepEqual(JSON.stringify(m), json);
   });
   
-  /*it('should stringify a Matrix containing a complex number, ccs storage format', function () {
+  it('should stringify a Matrix containing a complex number, ccs storage format', function () {
     var c = new Complex(4, 5);
     var m = math.matrix([[1,2],[3,c]], 'ccs');
     var json = '{"mathjs":"CcsMatrix","values":[1,3,2,{"mathjs":"Complex","re":4,"im":5}],"index":[0,1,0,1],"ptr":[0,2,4],"size":[2,2]}';
 
     assert.deepEqual(JSON.stringify(m), json);
-  });*/
+  });
 
   it('should stringify Help', function () {
     var h = new Help({name: 'foo', description: 'bar'});