+++ test/json/reviver.test.js
@@ -30,7 +30,7 @@ describe('reviver', function () {
 
   it('should parse a stringified BigNumber', function () {
     var json = '{"mathjs":"BigNumber","value":"0.2"}';
-    var b = new BigNumber(0.2);
+    var b = new math.type.BigNumber(0.2);
 
     var obj = JSON.parse(json, reviver);
 
@@ -120,6 +120,28 @@ describe('reviver', function () {
 
     assert(obj instanceof Matrix);
     assert(obj._data[1][1] instanceof Complex);
+    assert.deepEqual(obj, m);
+  });
+
+  it('should parse a Matrix, ccs storage format', function () {
+    var json = '{"mathjs":"CcsMatrix","values":[1,3,2,4],"index":[0,1,0,1],"ptr":[0,2,4],"size":[2,2]}';
+    var m = math.matrix([[1,2],[3,4]], 'ccs');
+
+    var obj = JSON.parse(json, reviver);
+
+    assert(obj instanceof math.type.CcsMatrix);
+    assert(obj instanceof Matrix);
+    assert.deepEqual(obj, m);
+  });
+
+  it('should parse a Matrix, crs storage format', function () {
+    var json = '{"mathjs":"CrsMatrix","values":[1,2,3,4],"index":[0,1,0,1],"ptr":[0,2,4],"size":[2,2]}';
+    var m = math.matrix([[1,2],[3,4]], 'crs');
+
+    var obj = JSON.parse(json, reviver);
+
+    assert(obj instanceof Matrix);
+    assert(obj instanceof math.type.CrsMatrix);
     assert.deepEqual(obj, m);
   });