+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## not yet released, version 1.4.1-SNAPSHOT
 
 - Added unit `stone` (6.35 kg).
+- Implemented BigNumber support for function `atan2`. Thanks @BigFav.
 - Improvements and bug fixes in outputting parentheses in `Node.toString` and
   `Node.toTex` functions. Thanks @FSMaxB.
 - Fixed #291: function `format` sometimes returning exponential notation when