+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 - Added unit `stone` (6.35 kg).
 - Implemented BigNumber support for function `atan2`. Thanks @BigFav.
+- Implemented support for custom LaTeX representations. Thanks @FSMaxB.
 - Improvements and bug fixes in outputting parentheses in `Node.toString` and
   `Node.toTex` functions. Thanks @FSMaxB.
 - Fixed #291: function `format` sometimes returning exponential notation when